move Data.Hie to Data.GHC.Hie
